Fix review round 2: abort controllers, drag cancel, reach integer
- Admin route: validate explicit task_priority_project_ids against
  effective overview-only set to prevent overlap
- Admin task rankings: add AbortController to useEffect fetch, log
  errors instead of silent catch, use s.userId as React key, remove
  IIFE in expanded view
- Task ranking view: add AbortController to load useEffect to prevent
  stale responses, add drag snapshot + onDragCancel handler for
  rollback on cancelled drags, add aria-label on drag handle
- Task RICE scoring view: add AbortController to load useEffect,
  reset state on project switch
- Migration: change reach column from numeric to integer to match
  API/UI contract

+++ supabase/migrations/20260312_hub_task_prioritisation.sql
@@ -40,7 +40,7 @@ create table if not exists hub_task_rice_scores (
   user_id text not null,
   project_linear_id text not null,
   issue_linear_id text not null,
-  reach numeric check (reach >= 1 and reach <= 10),
+  reach integer check (reach >= 1 and reach <= 10),
   impact numeric check (impact in (0.25, 0.5, 1, 2, 3)),
   confidence numeric check (confidence >= 0 and confidence <= 100),
   effort numeric check (effort >= 0.5),
